requirements.txt
google-generativeai
pandas
matplotlib
python-dotenv